Rename consumer test package to receiver
